+++ bindings/python/src/gar/expose-gar.cpp
@@ -14,8 +14,8 @@ using namespace gar;
 
 using context::Scalar;
 using riccati_base_t = RiccatiSolverBase<Scalar>;
-using knot_t = LQRKnotTpl<context::Scalar>;
-using lqr_t = LQRProblemTpl<context::Scalar>;
+using knot_t = LqrKnotTpl<context::Scalar>;
+using lqr_t = LqrProblemTpl<context::Scalar>;
 
 using context::MatrixXs;
 using RowMatrixXs = Eigen::Transpose<MatrixXs>::PlainMatrix;
@@ -58,9 +58,10 @@ void exposeGAR() {
 
   exposeBlockMatrices();
 
-  bp::class_<knot_t>("LQRKnot", bp::no_init)
-      .def(bp::init<uint, uint, uint>(("nx"_a, "nu", "nc")))
-      .def(bp::init<uint, uint, uint, uint>(("nx"_a, "nu", "nc", "nx2")))
+  bp::class_<knot_t>("LqrKnot", bp::no_init)
+      .def(bp::init<uint, uint, uint>(("self"_a, "nx", "nu", "nc")))
+      .def(bp::init<uint, uint, uint, uint, uint>(
+          ("self"_a, "nx"_a, "nu", "nc", "nx2", "nth"_a = 0)))
       .def_readonly("nx", &knot_t::nx)
       .def_readonly("nu", &knot_t::nu)
       .def_readonly("nc", &knot_t::nc)
@@ -93,9 +94,9 @@ void exposeGAR() {
       .def(CopyableVisitor<knot_t>())
       .def(PrintableVisitor<knot_t>());
 
-  StdVectorPythonVisitor<knot_vec_t, false>::expose("StdVec_LQRKnot");
+  StdVectorPythonVisitor<knot_vec_t, false>::expose("StdVec_LqrKnot");
 
-  bp::class_<lqr_t>("LQRProblem", bp::no_init)
+  bp::class_<lqr_t, boost::noncopyable>("LqrProblem", bp::no_init)
       .def(bp::init<const knot_vec_t &, long>(("self"_a, "stages", "nc0")))
       .def_readwrite("stages", &lqr_t::stages)
       .add_property("horizon", &lqr_t::horizon)
@@ -110,8 +111,7 @@ void exposeGAR() {
       .add_property("ntheta", &lqr_t::ntheta)
       .def("evaluate", &lqr_t::evaluate,
            ("self"_a, "xs", "us", "theta"_a = std::nullopt),
-           "Evaluate the problem objective.")
-      .def(CopyableVisitor<lqr_t>());
+           "Evaluate the problem objective.");
 
   bp::class_<riccati_base_t, boost::noncopyable>("RiccatiSolverBase",
                                                  bp::no_init)