+++ examples/gar-cycle-lqr-1d.py
@@ -14,7 +14,7 @@ nu = 1
 
 
 def create_knot(nx, nu):
-    knot = gar.LQRKnot(nx, nu, 0)
+    knot = gar.LqrKnot(nx, nu, 0)
     knot.Q[:] = np.eye(nx) * 0.01
     knot.R[:] = np.eye(nu) * 0.01
     knot.r[:] = (2 * np.random.rand(nx) - 1) * 0.01
@@ -29,7 +29,7 @@ T = 30
 base_knot = create_knot(nx, nu)
 knots = [base_knot for _ in range(T)]
 knots.append(create_knot(nx, 0))
-prob = gar.LQRProblem(knots, 0)
+prob = gar.LqrProblem(knots, 0)
 
 PARAM_DIM = nx
 prob.addParameterization(PARAM_DIM)