+++ gar/include/aligator/gar/proximal-riccati.hpp
@@ -1,7 +1,8 @@
+/// @copyright Copyright (C) 2023-2024 LAAS-CNRS, INRIA
 #pragma once
 
 #include "riccati-base.hpp"
-#include "riccati-impl.hpp"
+#include "riccati-kernel.hpp"
 
 namespace aligator {
 namespace gar {
@@ -20,9 +21,9 @@ public:
   using StageFactorType = StageFactor<Scalar>;
   using value_t = typename StageFactorType::value_t;
   using kkt0_t = typename Impl::kkt0_t;
-  using KnotType = LQRKnotTpl<Scalar>;
+  using KnotType = LqrKnotTpl<Scalar>;
 
-  explicit ProximalRiccatiSolver(const LQRProblemTpl<Scalar> &problem);
+  explicit ProximalRiccatiSolver(const LqrProblemTpl<Scalar> &problem);
 
   /// Backward sweep.
   bool backward(const Scalar mudyn, const Scalar mueq);
@@ -40,7 +41,7 @@ public:
   MatrixXs thHess; //< optimal value Hessian wrt parameter
 
 protected:
-  const LQRProblemTpl<Scalar> *problem_;
+  const LqrProblemTpl<Scalar> *problem_;
 };
 
 } // namespace gar