+++ gar/src/lqr-problem.cpp
@@ -3,7 +3,15 @@
 
 namespace aligator {
 namespace gar {
-template struct LQRKnotTpl<context::Scalar>;
-template struct LQRProblemTpl<context::Scalar>;
+template struct LqrKnotTpl<context::Scalar>;
+template struct LqrProblemTpl<context::Scalar>;
+
+static_assert(
+    std::uses_allocator_v<LqrKnotTpl<context::Scalar>, polymorphic_allocator>,
+    "");
+static_assert(std::uses_allocator_v<LqrProblemTpl<context::Scalar>,
+                                    polymorphic_allocator>,
+              "");
+
 } // namespace gar
 } // namespace aligator